refactor: 스트라이크 판정 로직 FallenFins 에 위임 및 패키지 정리
