--- src/main/java/bowling/BowlingGame.java
@@ -1,9 +1,0 @@
-package bowling;
-
-public class BowlingGame {
-
-    public static void main(String[] args) {
-        BowlingGameController.play();
-    }
-
-}
