+++ src/main/java/bowling/domain/ResultMark.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public enum ResultMark {
