+++ src/main/java/bowling/domain/exception/EndedFrameException.java
@@ -1,4 +1,4 @@
-package bowling.exception;
+package bowling.domain.exception;
 
 public class EndedFrameException extends RuntimeException {
