+++ src/main/java/bowling/domain/exception/ExceedFallenPinsException.java
@@ -1,4 +1,4 @@
-package bowling.exception;
+package bowling.domain.exception;
 
 public class ExceedFallenPinsException extends RuntimeException {
