+++ src/main/java/bowling/domain/exception/InvalidCountOfPinException.java
@@ -1,4 +1,4 @@
-package bowling.exception;
+package bowling.domain.exception;
 
 public class InvalidCountOfPinException extends RuntimeException {
