+++ src/test/java/bowling/domain/ResultMarkTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
