+++ src/test/java/bowling/domain/frame/BowlingFramesTest.java
@@ -1,4 +1,4 @@
-package bowling.frame;
+package bowling.domain.frame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
