+++ src/test/java/bowling/domain/frame/NormalFrameTest.java
@@ -1,11 +1,11 @@
-package bowling.frame;
+package bowling.domain.frame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
-import bowling.FallenPins;
-import bowling.exception.EndedFrameException;
-import bowling.exception.ExceedFallenPinsException;
+import bowling.domain.FallenPins;
+import bowling.domain.exception.EndedFrameException;
+import bowling.domain.exception.ExceedFallenPinsException;
 import org.junit.jupiter.api.Test;
 
 class NormalFrameTest {
@@ -13,7 +13,7 @@ class NormalFrameTest {
     @Test
     void 첫번째_투구에_모든_핀_쓰러뜨리면_스트라이크() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(10));
+        frame = frame.updateFrameState(FallenPins.of(10));
 
         assertThat(frame.getResult()).isEqualTo("X ");
     }
@@ -21,8 +21,8 @@ class NormalFrameTest {
     @Test
     void 첫번째_투구_후_핀이_남아있고_두번째_투구에_모든_핀_쓰러뜨리면_스페어() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(8));
-        frame = frame.update(FallenPins.of(2));
+        frame = frame.updateFrameState(FallenPins.of(8));
+        frame = frame.updateFrameState(FallenPins.of(2));
 
         assertThat(frame.getResult()).isEqualTo("8|/");
     }
@@ -30,8 +30,8 @@ class NormalFrameTest {
     @Test
     void 두번째_투구_후에도_모든_핀을_쓰러뜨리지_못하면_미스() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(8));
-        frame = frame.update(FallenPins.of(1));
+        frame = frame.updateFrameState(FallenPins.of(8));
+        frame = frame.updateFrameState(FallenPins.of(1));
 
         assertThat(frame.getResult()).isEqualTo("8|1");
     }
@@ -39,7 +39,7 @@ class NormalFrameTest {
     @Test
     void 첫번째_투구_시_하나도_핀을_못쓰러뜨리면_첫번째_거터() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(0));
+        frame = frame.updateFrameState(FallenPins.of(0));
 
         assertThat(frame.getResult()).isEqualTo("- ");
     }
@@ -47,8 +47,8 @@ class NormalFrameTest {
     @Test
     void 두번째_투구_시_하나도_핀을_못쓰러뜨리면_두번째_거터() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(7));
-        frame = frame.update(FallenPins.of(0));
+        frame = frame.updateFrameState(FallenPins.of(7));
+        frame = frame.updateFrameState(FallenPins.of(0));
 
         assertThat(frame.getResult()).isEqualTo("7|-");
     }
@@ -56,8 +56,8 @@ class NormalFrameTest {
     @Test
     void 첫번째_거터이며_두번째_투구_10개_전부_쓰러뜨리면_스페어() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(0));
-        frame = frame.update(FallenPins.of(10));
+        frame = frame.updateFrameState(FallenPins.of(0));
+        frame = frame.updateFrameState(FallenPins.of(10));
 
         assertThat(frame.getResult()).isEqualTo("-|/");
     }
@@ -65,7 +65,7 @@ class NormalFrameTest {
     @Test
     void 첫번째_투구_시_스트라이크는_아닐때_첫번째_점수_표기() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(4));
+        frame = frame.updateFrameState(FallenPins.of(4));
 
         assertThat(frame.getResult()).isEqualTo("4 ");
     }
@@ -73,8 +73,8 @@ class NormalFrameTest {
     @Test
     void 프레임_종료됨() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(7));
-        frame = frame.update(FallenPins.of(3));
+        frame = frame.updateFrameState(FallenPins.of(7));
+        frame = frame.updateFrameState(FallenPins.of(3));
 
         assertThat(frame.isFinish()).isTrue();
     }
@@ -82,7 +82,7 @@ class NormalFrameTest {
     @Test
     void 프레임_종료됨_스트라이크() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(10));
+        frame = frame.updateFrameState(FallenPins.of(10));
 
         assertThat(frame.isFinish()).isTrue();
     }
@@ -90,7 +90,7 @@ class NormalFrameTest {
     @Test
     void 프레임_진행중() {
         Frame frame = new NormalFrame();
-        frame = frame.update(FallenPins.of(7));
+        frame = frame.updateFrameState(FallenPins.of(7));
 
         assertThat(frame.isFinish()).isFalse();
     }
@@ -100,8 +100,8 @@ class NormalFrameTest {
         assertThatExceptionOfType(ExceedFallenPinsException.class)
                 .isThrownBy(() -> {
                     NormalFrame frame = new NormalFrame();
-                    frame = frame.update(FallenPins.of(7));
-                    frame = frame.update(FallenPins.of(7));
+                    frame = frame.updateFrameState(FallenPins.of(7));
+                    frame = frame.updateFrameState(FallenPins.of(7));
                 });
     }
 
@@ -110,8 +110,8 @@ class NormalFrameTest {
         assertThatExceptionOfType(EndedFrameException.class)
                 .isThrownBy(() -> {
                     NormalFrame frame = new NormalFrame();
-                    frame = frame.update(FallenPins.of(10));
-                    frame = frame.update(FallenPins.of(0));
+                    frame = frame.updateFrameState(FallenPins.of(10));
+                    frame = frame.updateFrameState(FallenPins.of(0));
                 });
     }
 
@@ -120,9 +120,9 @@ class NormalFrameTest {
         assertThatExceptionOfType(EndedFrameException.class)
                 .isThrownBy(() -> {
                     NormalFrame frame = new NormalFrame();
-                    frame = frame.update(FallenPins.of(0));
-                    frame = frame.update(FallenPins.of(10));
-                    frame = frame.update(FallenPins.of(4));
+                    frame = frame.updateFrameState(FallenPins.of(0));
+                    frame = frame.updateFrameState(FallenPins.of(10));
+                    frame = frame.updateFrameState(FallenPins.of(4));
                 });
     }
